add a autocomplete utility
